echoguard/requirements.txt
# EchoGuard - Python dependencies (exam-safe, student project)
# Use only: pip install -r requirements.txt

flask>=3.0.0
google-genai>=1.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
pypdf>=3.0.0
